Final Push  - Ready for Review (2)
Added the speech video to the page. The speech has the two possible formats, and has a poster that appears while it is loading. Added the captions to the video, and a text that appears should the video not. Added CSS to make the video take up less space and be in the top right of the first column. The captions now have a gray background with larger text and a pink sans serif font. Added Comments.